email-scan connections refactored
