+++ api-js/src/email-scan/objects/index.js
@@ -1,5 +1,9 @@
 export * from './dkim'
+export * from './dkim-connection'
 export * from './dkim-result'
+export * from './dkim-result-connection'
 export * from './dmarc'
+export * from './dmarc-connection'
 export * from './email-scan'
 export * from './spf'
+export * from './spf-connection'
